+++ RELEASES.md
@@ -12,6 +12,8 @@ The TODO section is part of the [README.md](https://github.com/automation-tasks-
 
 - upload
 
+- snake sprite
+
 ## Version 0.0.74 (2025-07-20)
 
 - update auto